snowflake/requirements.txt
snowflake-connector-python==3.5.0
pandas==2.1.4
supabase==2.3.0
python-dotenv==1.0.0
pyarrow==14.0.1